Mock lister: move utils to uti.go, rm useless mocks
