+++ test/unit/listers/BUILD.bazel
@@ -6,6 +6,7 @@ go_library(
         "certificaterequest.go",
         "order.go",
         "secret.go",
+        "util.go",
     ],
     importpath = "github.com/jetstack/cert-manager/test/unit/listers",
     visibility = ["//visibility:public"],
